requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.42
psycopg2-binary==2.9.10
python-jose==3.5.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.13.1
pydantic==2.10.4
python-dotenv==1.0.1